Add EditingIndicator and UploadStatusIndicator components; enhance FilePreview and EditHistoryTooltip functionality
